abstract the class 'AbstractEventParser'
