+++ parse/src/main/java/com/alibaba/otter/canal/parse/ParserMetrics.java
@@ -1,0 +1,12 @@
+package com.alibaba.otter.canal.parse;
+
+import java.util.concurrent.atomic.AtomicLong;
+
+public interface ParserMetrics {
+
+    boolean isParallel();
+
+    AtomicLong getEventsPublishBlockingTime();
+
+    AtomicLong getReceivedBinlogBytes();
+}
