+++ parse/src/main/java/com/alibaba/otter/canal/parse/inbound/mysql/AbstractMysqlEventParser.java
@@ -1,26 +1,40 @@
 package com.alibaba.otter.canal.parse.inbound.mysql;
 
+import java.io.IOException;
 import java.nio.charset.Charset;
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.Arrays;
+import java.util.Timer;
+import java.util.TimerTask;
+import java.util.concurrent.atomic.AtomicBoolean;
+
+import org.apache.commons.lang.StringUtils;
 
 import com.alibaba.otter.canal.filter.CanalEventFilter;
 import com.alibaba.otter.canal.filter.aviater.AviaterRegexFilter;
 import com.alibaba.otter.canal.parse.CanalEventParser;
+import com.alibaba.otter.canal.parse.driver.mysql.packets.GTIDSet;
 import com.alibaba.otter.canal.parse.exception.CanalParseException;
+import com.alibaba.otter.canal.parse.exception.PositionNotFoundException;
 import com.alibaba.otter.canal.parse.inbound.AbstractEventParser;
+import com.alibaba.otter.canal.parse.inbound.BinlogConnection;
 import com.alibaba.otter.canal.parse.inbound.BinlogParser;
+import com.alibaba.otter.canal.parse.inbound.ErosaConnection;
 import com.alibaba.otter.canal.parse.inbound.MultiStageCoprocessor;
+import com.alibaba.otter.canal.parse.inbound.SinkFunction;
 import com.alibaba.otter.canal.parse.inbound.mysql.dbsync.LogEventConvert;
 import com.alibaba.otter.canal.parse.inbound.mysql.tsdb.DatabaseTableMeta;
 import com.alibaba.otter.canal.parse.inbound.mysql.tsdb.DefaultTableMetaTSDBFactory;
 import com.alibaba.otter.canal.parse.inbound.mysql.tsdb.TableMetaTSDB;
 import com.alibaba.otter.canal.parse.inbound.mysql.tsdb.TableMetaTSDBFactory;
+import com.alibaba.otter.canal.protocol.CanalEntry;
 import com.alibaba.otter.canal.protocol.position.EntryPosition;
-import org.apache.commons.lang.StringUtils;
+import com.taobao.tddl.dbsync.binlog.exception.TableIdNotFoundException;
+
+import static com.alibaba.otter.canal.parse.driver.mysql.utils.GtidUtil.parseGtidSet;
 
 public abstract class AbstractMysqlEventParser extends AbstractEventParser {
 
-    protected static final long    BINLOG_START_OFFEST       = 4L;
+    protected static final long    BINLOG_START_OFFSET = 4L;
 
     protected TableMetaTSDBFactory tableMetaTSDBFactory      = new DefaultTableMetaTSDBFactory();
     protected boolean              enableTsdb                = false;
@@ -42,10 +56,21 @@ public abstract class AbstractMysqlEventParser extends AbstractEventParser {
     protected boolean              filterDmlInsert           = false;
     protected boolean              filterDmlUpdate           = false;
     protected boolean              filterDmlDelete           = false;
-    // instance received binlog bytes
-    protected final AtomicLong     receivedBinlogBytes       = new AtomicLong(0L);
-    private final AtomicLong       eventsPublishBlockingTime = new AtomicLong(0L);
 
+    protected AtomicBoolean        needTransactionPosition    = new AtomicBoolean(false);
+    protected volatile boolean     detectingEnable            = true;                                    // 是否开启心跳检查
+    protected Integer              detectingIntervalInSeconds = 3;                                       // 检测频率
+    protected volatile Timer       timer;
+    protected TimerTask            heartBeatTimerTask;
+
+    protected boolean              isGTIDMode                 = false;                                   // 是否是GTID模式
+    protected long                 serverId;
+
+    protected abstract ErosaConnection buildErosaConnection();
+
+    protected abstract EntryPosition findStartPosition(ErosaConnection connection) throws IOException;
+
+    @Override
     protected BinlogParser buildParser() {
         LogEventConvert convert = new LogEventConvert();
         if (eventFilter != null && eventFilter instanceof AviaterRegexFilter) {
@@ -69,6 +94,162 @@ public abstract class AbstractMysqlEventParser extends AbstractEventParser {
         return convert;
     }
 
+    @Override
+    protected BinlogConnection buildBinlogConnection() {
+        return buildErosaConnection();
+    }
+
+    @Override
+    protected void dump(BinlogConnection connection) throws IOException {
+        ErosaConnection erosaConnection = (ErosaConnection) connection;
+        boolean isMariaDB = false;
+        try {
+            // 启动一个心跳线程
+            startHeartBeat(erosaConnection);
+            // 连接
+            erosaConnection.connect();
+
+            long queryServerId = erosaConnection.queryServerId();
+            if (queryServerId != 0) {
+                serverId = queryServerId;
+            }
+
+            if (erosaConnection instanceof MysqlConnection) {
+                isMariaDB = ((MysqlConnection)erosaConnection).isMariaDB();
+            }
+            // 4. 获取最后的位置信息
+            long start = System.currentTimeMillis();
+            logger.warn("---> begin to find start position, it will be long time for reset or first position");
+            EntryPosition position = findStartPosition(erosaConnection);
+            final EntryPosition startPosition = position;
+            if (startPosition == null) {
+                throw new PositionNotFoundException("can't find start position for " + destination);
+            }
+
+            if (!processTableMeta(startPosition)) {
+                throw new CanalParseException("can't find init table meta for " + destination
+                                              + " with position : " + startPosition);
+            }
+            long end = System.currentTimeMillis();
+            logger.warn("---> find start position successfully, {}", startPosition + " cost : "
+                                                                     + (end - start)
+                                                                     + "ms , the next step is binlog dump");
+            // 重新链接，因为在找position过程中可能有状态，需要断开后重建
+            erosaConnection.reconnect();
+
+            final SinkFunction sinkHandler = buildSinkHandler(startPosition);
+
+            // 4. 开始dump数据
+            if (parallel) {
+                // build stage processor
+                multiStageCoprocessor = buildMultiStageCoprocessor();
+                if (isGTIDMode() && StringUtils.isNotEmpty(startPosition.getGtid())) {
+                    // 判断所属instance是否启用GTID模式，是的话调用ErosaConnection中GTID对应方法dump数据
+                    GTIDSet gtidSet = parseGtidSet(startPosition.getGtid(),isMariaDB);
+                    ((MysqlMultiStageCoprocessor) multiStageCoprocessor).setGtidSet(gtidSet);
+                    multiStageCoprocessor.start();
+                    erosaConnection.dump(gtidSet, multiStageCoprocessor);
+                } else {
+                    multiStageCoprocessor.start();
+                    if (StringUtils.isEmpty(startPosition.getJournalName())
+                        && startPosition.getTimestamp() != null) {
+                        erosaConnection.dump(startPosition.getTimestamp(), multiStageCoprocessor);
+                    } else {
+                        erosaConnection.dump(startPosition.getJournalName(),
+                                startPosition.getPosition(),
+                                multiStageCoprocessor);
+                    }
+                }
+            } else {
+                if (isGTIDMode() && StringUtils.isNotEmpty(startPosition.getGtid())) {
+                    // 判断所属instance是否启用GTID模式，是的话调用ErosaConnection中GTID对应方法dump数据
+                    erosaConnection.dump(parseGtidSet(startPosition.getGtid(), isMariaDB), sinkHandler);
+                } else {
+                    if (StringUtils.isEmpty(startPosition.getJournalName())
+                        && startPosition.getTimestamp() != null) {
+                        erosaConnection.dump(startPosition.getTimestamp(), sinkHandler);
+                    } else {
+                        erosaConnection.dump(startPosition.getJournalName(),
+                                startPosition.getPosition(),
+                                sinkHandler);
+                    }
+                }
+            }
+        } catch (TableIdNotFoundException e) {
+            exception = e;
+            // 特殊处理TableIdNotFound异常,出现这样的异常，一种可能就是起始的position是一个事务当中，导致tablemap
+            // Event时间没解析过
+            needTransactionPosition.compareAndSet(false, true);
+            logger.error(String.format("dump address %s has an error, retrying. caused by ",
+                    runningInfo.getAddress().toString()), e);
+        }
+    }
+
+    protected void startHeartBeat(ErosaConnection connection) {
+        lastEntryTime = 0L; // 初始化
+        if (timer == null) {// lazy初始化一下
+            String name = String.format("destination = %s , address = %s , HeartBeatTimeTask",
+                    destination,
+                    runningInfo == null ? null : runningInfo.getAddress().toString());
+            synchronized (AbstractEventParser.class) {
+                // synchronized (MysqlEventParser.class) {
+                // why use MysqlEventParser.class, u know, MysqlEventParser is
+                // the child class 4 AbstractEventParser,
+                // do this is ...
+                if (timer == null) {
+                    timer = new Timer(name, true);
+                }
+            }
+        }
+
+        if (heartBeatTimerTask == null) {// fixed issue #56，避免重复创建heartbeat线程
+            heartBeatTimerTask = buildHeartBeatTimeTask(connection);
+            Integer interval = detectingIntervalInSeconds;
+            timer.schedule(heartBeatTimerTask, interval * 1000L, interval * 1000L);
+            logger.info("start heart beat.... ");
+        }
+    }
+
+    protected TimerTask buildHeartBeatTimeTask(ErosaConnection connection) {
+        return new TimerTask() {
+
+            @Override
+            public void run() {
+                try {
+                    if (exception == null || lastEntryTime > 0) {
+                        // 如果未出现异常，或者有第一条正常数据
+                        long now = System.currentTimeMillis();
+                        long inteval = (now - lastEntryTime) / 1000;
+                        if (inteval >= detectingIntervalInSeconds) {
+                            CanalEntry.Header.Builder headerBuilder = CanalEntry.Header.newBuilder();
+                            headerBuilder.setExecuteTime(now);
+                            CanalEntry.Entry.Builder entryBuilder = CanalEntry.Entry.newBuilder();
+                            entryBuilder.setHeader(headerBuilder.build());
+                            entryBuilder.setEntryType(CanalEntry.EntryType.HEARTBEAT);
+                            CanalEntry.Entry entry = entryBuilder.build();
+                            // 提交到sink中，目前不会提交到store中，会在sink中进行忽略
+                            consumeTheEventAndProfilingIfNecessary(Arrays.asList(entry));
+                        }
+                    }
+
+                } catch (Throwable e) {
+                    logger.warn("heartBeat run failed ", e);
+                }
+            }
+
+        };
+    }
+
+    protected void stopHeartBeat() {
+        lastEntryTime = 0L; // 初始化
+        if (timer != null) {
+            timer.cancel();
+            timer = null;
+        }
+        heartBeatTimerTask = null;
+    }
+
+    @Override
     public void setEventFilter(CanalEventFilter eventFilter) {
         super.setEventFilter(eventFilter);
 
@@ -84,6 +265,7 @@ public abstract class AbstractMysqlEventParser extends AbstractEventParser {
         }
     }
 
+    @Override
     public void setEventBlackFilter(CanalEventFilter eventBlackFilter) {
         super.setEventBlackFilter(eventBlackFilter);
 
@@ -145,6 +327,7 @@ public abstract class AbstractMysqlEventParser extends AbstractEventParser {
         return true;
     }
 
+    @Override
     public void start() throws CanalParseException {
         if (enableTsdb) {
             if (tableMetaTSDB == null) {
@@ -164,15 +347,17 @@ public abstract class AbstractMysqlEventParser extends AbstractEventParser {
         super.start();
     }
 
+    @Override
     public void stop() throws CanalParseException {
         if (enableTsdb) {
             tableMetaTSDBFactory.destory(destination);
             tableMetaTSDB = null;
         }
-
+        stopHeartBeat();
         super.stop();
     }
 
+    @Override
     protected MultiStageCoprocessor buildMultiStageCoprocessor() {
         MysqlMultiStageCoprocessor mysqlMultiStageCoprocessor = new MysqlMultiStageCoprocessor(parallelBufferSize,
             parallelThreadSize,
@@ -277,14 +462,6 @@ public abstract class AbstractMysqlEventParser extends AbstractEventParser {
         this.tableMetaTSDBFactory = tableMetaTSDBFactory;
     }
 
-    public AtomicLong getEventsPublishBlockingTime() {
-        return this.eventsPublishBlockingTime;
-    }
-
-    public AtomicLong getReceivedBinlogBytes() {
-        return this.receivedBinlogBytes;
-    }
-
     public int getTsdbSnapshotInterval() {
         return tsdbSnapshotInterval;
     }
@@ -301,4 +478,27 @@ public abstract class AbstractMysqlEventParser extends AbstractEventParser {
         this.tsdbSnapshotExpire = tsdbSnapshotExpire;
     }
 
+    public void setDetectingEnable(boolean detectingEnable) {
+        this.detectingEnable = detectingEnable;
+    }
+
+    public void setDetectingIntervalInSeconds(Integer detectingIntervalInSeconds) {
+        this.detectingIntervalInSeconds = detectingIntervalInSeconds;
+    }
+
+    public boolean isGTIDMode() {
+        return isGTIDMode;
+    }
+
+    public void setIsGTIDMode(boolean isGTIDMode) {
+        this.isGTIDMode = isGTIDMode;
+    }
+
+    public long getServerId() {
+        return serverId;
+    }
+
+    public void setServerId(long serverId) {
+        this.serverId = serverId;
+    }
 }
