+++ parse/src/main/java/com/alibaba/otter/canal/parse/inbound/mysql/LocalBinlogEventParser.java
@@ -6,11 +6,11 @@ import org.apache.commons.lang.StringUtils;
 
 import com.alibaba.otter.canal.parse.CanalEventParser;
 import com.alibaba.otter.canal.parse.exception.CanalParseException;
+import com.alibaba.otter.canal.parse.inbound.BinlogConnection;
 import com.alibaba.otter.canal.parse.inbound.ErosaConnection;
 import com.alibaba.otter.canal.parse.inbound.mysql.dbsync.LogEventConvert;
 import com.alibaba.otter.canal.parse.inbound.mysql.dbsync.TableMetaCache;
 import com.alibaba.otter.canal.parse.inbound.mysql.tsdb.DatabaseTableMeta;
-import com.alibaba.otter.canal.parse.index.CanalLogPositionManager;
 import com.alibaba.otter.canal.parse.support.AuthenticationInfo;
 import com.alibaba.otter.canal.protocol.position.EntryPosition;
 import com.alibaba.otter.canal.protocol.position.LogPosition;
@@ -43,7 +43,7 @@ public class LocalBinlogEventParser extends AbstractMysqlEventParser implements 
     }
 
     @Override
-    protected void preDump(ErosaConnection connection) {
+    protected void preDump(BinlogConnection connection) {
         metaConnection = buildMysqlConnection();
         try {
             metaConnection.connect();
@@ -65,7 +65,7 @@ public class LocalBinlogEventParser extends AbstractMysqlEventParser implements 
     }
 
     @Override
-    protected void afterDump(ErosaConnection connection) {
+    protected void afterDump(BinlogConnection connection) {
         if (metaConnection != null) {
             try {
                 metaConnection.disconnect();
@@ -76,6 +76,7 @@ public class LocalBinlogEventParser extends AbstractMysqlEventParser implements 
         }
     }
 
+    @Override
     public void start() throws CanalParseException {
         if (runningInfo == null) { // 第一次链接主库
             runningInfo = masterInfo;
@@ -159,10 +160,6 @@ public class LocalBinlogEventParser extends AbstractMysqlEventParser implements 
     }
 
     // ========================= setter / getter =========================
-
-    public void setLogPositionManager(CanalLogPositionManager logPositionManager) {
-        this.logPositionManager = logPositionManager;
-    }
 
     public void setDirectory(String directory) {
         this.directory = directory;
